tools: move shared python functions to common.py
No functional change, only move code here and there,
and add some comment.
